性能优化批B：时间精度、帧内分配、EventBus/ActionDatabase/GetSystem
- Timestep 帧差改为 double 计算再窄化（float 累计 77 秒后丢毫秒精度），
  m_LastFrameTime 改 double
- RenderSystem/UISystem 的每帧 sort vector 提为成员复用，消除每帧堆分配
- EventBus：订阅时保持 handler 有序，Dispatch 用 inplace_merge 归并两段
  已排序序列（O(n)），不再每事件 std::sort
- ActionDatabase 从 vector+线性查找改为 unordered_map（Find/Has/Register
  O(1)，All() 遍历收集保持 API）
- Scene::GetSystem 从每帧 linear dynamic_cast 扫描改为 type_index 缓存
  （RegisterSystem 时标脏，惰性重建）

+++ Wheatear/src/Wheatear/Core/Application.h
@@ -104,7 +104,7 @@ namespace Wheatear {
 		std::vector<Layer*> m_PendingOverlaysToPush;
 		std::vector<Layer*> m_PendingOverlaysToPop;
 
-		float m_LastFrameTime = 0.0f;
+		double m_LastFrameTime = 0.0;
 
 		static Application* s_Instance;
 	};
